03_quant_trading/01_hft_microstructure_alpha_signals/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
numba>=0.57.0
joblib>=1.3.0
tqdm>=4.65.0
